Update main
Menu Update: Additional menu option was added to restrict the Tree Depth based on the user decision.

4.A and 4.B Update: This section of the code will now take into consideration of the user tree depth restriction. Figures that are displayed will now have a title and number count on which figure of the recursion that we are looking at.

Please note that the restricted depth tree that the user chooses does not affect the processing of the data as this is only for a visualization purpose. as stated in the instruction.

README File: ReadMe file has also been updated for further instructions for the restriction tree depth.